3.DRF_API/4주차/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
idna==2.10
importlib-metadata==3.7.2
importlib-resources==5.1.2
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
zipp==3.4.1
